split App.tsx into utils modules (cleaner)
